Refactor seeds and delete configs
