--- src/openelm/config/benchmark_cfg_triton.yaml
@@ -1,21 +1,0 @@
-triton_host: ???
-triton_port: 8001
-top_p: 0.9
-temperature: 0.8
-timeout: 5  # Seconds
-gen_max_length: 128
-batch_size: 32
-n_trials: 3200
-n_bugs: 1
-fp16: False
-pad_token: 50256
-tasks: ["parity"]
-run_name: ???  # Mandatory string argument that describes the run.
-
-###################################################################################################
-# Hydra config overrides:
-hydra:
-    run:
-        dir: logs/${run_name}
-    sweep:
-        dir: logs/${run_name}
