--- src/openelm/config/elm_sodarace_cfg.yaml
@@ -1,27 +1,0 @@
-model: codegen-350M-mono
-cuda: True
-gpus: 1
-seed: 42
-deterministic: False
-fp16: False
-top_p: 0.95
-temp: 0.85
-timeout: 5.0  # Seconds
-gen_max_len: 768
-batch_size: 32
-evo_init_steps: 10
-evo_n_steps: 20
-behavior_n_bins: 12
-evo_history_length: 10
-evaluation_steps: 1000
-pad_token: 50256
-env_name: "sodarace"
-run_name: ???  # Mandatory string argument that describes the run.
-
-###################################################################################################
-# Hydra config overrides:
-hydra:
-    run:
-        dir: logs/${run_name}
-    sweep:
-        dir: logs/${run_name}
